feature, doc: slight dimensions adjusments (++ Hidden dim to follow ViT), rename doc folder
