+++ source/run_experiment.py
@@ -18,19 +18,19 @@ from source.model.blocks.constants.files import *
 
 import torch
 
-ENCODER_HEADS = 6
-DECODER_HEADS = 6
+ENCODER_HEADS = 16
+DECODER_HEADS = 16
 
-ENCODER_LAYERS = 8
-DECODER_LAYERS = 8
+ENCODER_LAYERS = 12
+DECODER_LAYERS = 12
 
 AUTOREGRESS_TARGET_LEN = 500
 
 DROPOUT_RATIO = 0.1
 
-BATCH_SIZE = 8
-PATCHES_DIM = (4, 4)
-EMBEDDING_DIMS = 12
+BATCH_SIZE = 16
+PATCHES_DIM = (8, 8)
+EMBEDDING_DIMS = 256
 
 NORMALIZE_COORDS = True
 NORMALIZE_PIXEL_VALUES = True
@@ -44,6 +44,8 @@ TEST_SIZE = 0.2
 LR = 0.001
 N_EPOCHS = 5
 
+USE_BRUSH = True
+
 from source.model.blocks.constants.device_helper import device
 
 if __name__ == "__main__":
@@ -52,9 +54,9 @@ if __name__ == "__main__":
     #     logger.add_log_channel(channel)
     #logger.add_log_channel(LogChannels.TRAINING)
     #logger.add_log_channel(LogChannels.DEBUG)
-    #logger.add_log_channel(LogChannels.LOSSES)
+    logger.add_log_channel(LogChannels.LOSSES)
     # logger.add_log_channel(LogChannels.INIT)
-    # logger.add_log_channel(LogChannels.PARAMS)
+    logger.add_log_channel(LogChannels.PARAMS)
     # logger.add_log_channel(LogChannels.DIMENSIONS)
     logger.add_log_channel(LogChannels.DATA)
     # logger.add_log_channel(LogChannels.PADDING)
@@ -63,14 +65,15 @@ if __name__ == "__main__":
     #print(f"Using device: {device} ({torch.cuda.get_device_name(device) if torch.cuda.is_available() else ''})")
 
     #Init data management
-    # dataset = BrushDataset(brush_root=BRUSH_ROOT, patches_dim=PATCHES_DIM, save_to_file=False, 
-    #                        strokemode=True, window_size=0,
-    #                        normalize_pixel_values=NORMALIZE_PIXEL_VALUES, normalize_coordinate_sequences=NORMALIZE_COORDS)
-    
-    dataset = UnipenDataset(unipen_root=UNIPEN_ROOT, patches_dim=PATCHES_DIM,
-                           strokemode=True, window_size=0,
-                           normalize_pixel_values=NORMALIZE_PIXEL_VALUES, normalize_coordinate_sequences=NORMALIZE_COORDS,
-                           samples_to_take=5000)
+    if USE_BRUSH:
+        dataset = BrushDataset(brush_root=BRUSH_ROOT, patches_dim=PATCHES_DIM, save_to_file=False, 
+                            strokemode=True, window_size=0,
+                            normalize_pixel_values=NORMALIZE_PIXEL_VALUES, normalize_coordinate_sequences=NORMALIZE_COORDS)
+    else:
+        dataset = UnipenDataset(unipen_root=UNIPEN_ROOT, patches_dim=PATCHES_DIM,
+                            strokemode=True, window_size=0,
+                            normalize_pixel_values=NORMALIZE_PIXEL_VALUES, normalize_coordinate_sequences=NORMALIZE_COORDS,
+                            samples_to_take=5000)
 
     dataset.transform_to_batch()
